requirements.txt
Django==5.0.6
tronpy==0.5.0
djangorestframework==3.15.0
python-dotenv==1.0.1